feat: let LSP suggest trait impl methods as you are typing them
