+++ tooling/lsp/src/requests/completion.rs
@@ -4,7 +4,10 @@ use std::{
 };
 
 use async_lsp::ResponseError;
-use completion_items::{field_completion_item, simple_completion_item, snippet_completion_item};
+use completion_items::{
+    field_completion_item, simple_completion_item, snippet_completion_item,
+    trait_impl_method_completion_item,
+};
 use convert_case::{Case, Casing};
 use fm::{FileId, FileMap, PathString};
 use kinds::{FunctionCompletionKind, FunctionKind, RequestedItems};
@@ -15,8 +18,9 @@ use noirc_frontend::{
         AsTraitPath, AttributeTarget, BlockExpression, CallExpression, ConstructorExpression,
         Expression, ExpressionKind, ForLoopStatement, GenericTypeArgs, Ident, IfExpression,
         ItemVisibility, Lambda, LetStatement, MemberAccessExpression, MethodCallExpression,
-        NoirFunction, NoirStruct, NoirTraitImpl, Path, PathKind, Pattern, Statement, TypeImpl,
-        UnresolvedGeneric, UnresolvedGenerics, UnresolvedType, UseTree, UseTreeKind, Visitor,
+        NoirFunction, NoirStruct, NoirTraitImpl, Path, PathKind, Pattern, Statement,
+        TraitImplItemKind, TypeImpl, UnresolvedGeneric, UnresolvedGenerics, UnresolvedType,
+        UseTree, UseTreeKind, Visitor,
     },
     graph::{CrateId, Dependency},
     hir::def_map::{CrateDefMap, LocalModuleId, ModuleId},
@@ -29,7 +33,10 @@ use noirc_frontend::{
 };
 use sort_text::underscore_sort_text;
 
-use crate::{requests::to_lsp_location, utils, visibility::is_visible, LspState};
+use crate::{
+    requests::to_lsp_location, trait_impl_method_stub_generator::TraitImplMethodStubGenerator,
+    utils, visibility::is_visible, LspState,
+};
 
 use super::process_request;
 
@@ -81,6 +88,7 @@ pub(crate) fn on_completion_request(
 struct NodeFinder<'a> {
     files: &'a FileMap,
     file: FileId,
+    source: &'a str,
     lines: Vec<&'a str>,
     byte_index: usize,
     byte: Option<u8>,
@@ -133,6 +141,7 @@ impl<'a> NodeFinder<'a> {
         Self {
             files,
             file,
+            source,
             lines: source.lines().collect(),
             byte_index,
             byte,
@@ -843,6 +852,73 @@ impl<'a> NodeFinder<'a> {
         }
     }
 
+    fn suggest_trait_impl_function(
+        &mut self,
+        noir_trait_impl: &NoirTraitImpl,
+        noir_function: &NoirFunction,
+    ) {
+        // First find the trait
+        let location = Location::new(noir_trait_impl.trait_name.span(), self.file);
+        let Some(ReferenceId::Trait(trait_id)) = self.interner.find_referenced(location) else {
+            return;
+        };
+
+        let trait_ = self.interner.get_trait(trait_id);
+
+        // Get all methods
+        let mut method_ids = trait_.method_ids.clone();
+
+        // Remove the ones that already are implemented
+        for item in &noir_trait_impl.items {
+            if let TraitImplItemKind::Function(noir_function) = &item.item.kind {
+                method_ids.remove(noir_function.name());
+            }
+        }
+
+        let indent = 0;
+
+        // Suggest the ones that match the name
+        let prefix = noir_function.name();
+        for (name, func_id) in method_ids {
+            if !name_matches(&name, prefix) {
+                continue;
+            }
+
+            let func_meta = self.interner.function_meta(&func_id);
+
+            let mut generator = TraitImplMethodStubGenerator::new(
+                &name,
+                func_meta,
+                trait_,
+                noir_trait_impl,
+                self.interner,
+                self.def_maps,
+                self.module_id,
+                indent,
+            );
+            generator.set_body("${1}".to_string());
+
+            let stub = generator.generate();
+
+            // We don't need the initial indent nor the final newlines
+            let stub = stub.trim();
+            // We also don't need the leading "fn " as that's already in the code;
+            let stub = stub.strip_prefix("fn ").unwrap();
+
+            let label = if func_meta.parameters.is_empty() {
+                format!("fn {}()", &name)
+            } else {
+                format!("fn {}(..)", &name)
+            };
+
+            let completion_item = trait_impl_method_completion_item(label, stub);
+            let completion_item = self
+                .completion_item_with_doc_comments(ReferenceId::Function(func_id), completion_item);
+
+            self.completion_items.push(completion_item);
+        }
+    }
+
     fn try_set_self_type(&mut self, pattern: &Pattern) {
         match pattern {
             Pattern::Identifier(ident) => {
@@ -949,6 +1025,24 @@ impl<'a> Visitor for NodeFinder<'a> {
         self.collect_type_parameters_in_generics(&noir_trait_impl.impl_generics);
 
         for item in &noir_trait_impl.items {
+            if let TraitImplItemKind::Function(noir_function) = &item.item.kind {
+                // Check if it's `fn foo>|<` and neither `(` nor `<` follow
+                if noir_function.name_ident().span().end() as usize == self.byte_index
+                    && noir_function.parameters().is_empty()
+                {
+                    let bytes = self.source.as_bytes();
+                    let mut cursor = self.byte_index;
+                    while cursor < bytes.len() && bytes[cursor].is_ascii_whitespace() {
+                        cursor += 1;
+                    }
+                    let char = bytes[cursor] as char;
+                    if char != '(' && char != '<' {
+                        self.suggest_trait_impl_function(noir_trait_impl, noir_function);
+                        return false;
+                    }
+                }
+            }
+
             item.item.accept(self);
         }
